Split up native interface code into files
